render rects in one draw call
