+++ alacritty/res/rect.f.glsl
@@ -14,7 +14,7 @@
 #version 300 es
 precision mediump float;
 
-uniform vec4 color;
+flat in vec4 color;
 
 out vec4 FragColor;
