+++ alacritty/res/rect.v.glsl
@@ -13,8 +13,12 @@
 // limitations under the License.
 #version 300 es
 layout (location = 0) in vec2 aPos;
+layout (location = 1) in vec4 aColor;
+
+flat out vec4 color;
 
 void main()
 {
+		color = aColor;
     gl_Position = vec4(aPos.x, aPos.y, 0.0, 1.0);
 }
