+++ alacritty/src/renderer/solidrect.rs
@@ -1,0 +1,203 @@
+use super::math::*;
+use super::rects::RenderRect;
+use super::shade::RectShaderProgram;
+use super::Glyph;
+use crate::gl;
+use crate::gl::types::*;
+use crate::renderer::Error;
+use alacritty_terminal::term::color::Rgb;
+use alacritty_terminal::term::SizeInfo;
+
+use std::mem::size_of;
+use std::ptr;
+
+pub enum RectAddError {
+    Full,
+}
+
+#[repr(C)]
+#[derive(Debug, Clone, Copy)]
+struct Rgba {
+    r: u8,
+    g: u8,
+    b: u8,
+    a: u8,
+}
+
+#[repr(C)]
+#[derive(Debug, Clone, Copy)]
+struct Vertex {
+    x: f32,
+    y: f32,
+    color: Rgba,
+}
+
+#[derive(Debug)]
+pub struct Rectifier {
+    vao: GLuint,
+    vbo: GLuint,
+    ebo: GLuint,
+    program: RectShaderProgram,
+    indices: Vec<u16>,
+    vertices: Vec<Vertex>,
+    size_info: Option<SizeInfo>,
+}
+
+impl Rectifier {
+    pub fn new() -> Result<Self, Error> {
+        let mut vao: GLuint = 0;
+        let mut vbo: GLuint = 0;
+        let mut ebo: GLuint = 0;
+
+        unsafe {
+            gl::GenVertexArrays(1, &mut vao);
+            gl::GenBuffers(1, &mut vbo);
+            gl::GenBuffers(1, &mut ebo);
+        }
+
+        Ok(Self {
+            vao,
+            vbo,
+            ebo,
+            program: RectShaderProgram::new()?,
+            indices: Vec::new(),
+            vertices: Vec::new(),
+            size_info: None,
+        })
+    }
+
+    pub fn begin(&mut self, size_info: &SizeInfo) {
+        self.indices.clear();
+        self.vertices.clear();
+        self.size_info = Some(*size_info);
+    }
+
+    pub fn add(&mut self, rect: &RenderRect) -> Result<(), RectAddError> {
+        let index = self.vertices.len();
+        if index >= 65536 - 4 {
+            return Err(RectAddError::Full);
+        }
+        let index = index as u16;
+
+        if rect.alpha <= 0. {
+            return Ok(());
+        }
+
+        let size_info = self.size_info.as_ref().unwrap();
+
+        // Calculate rectangle position.
+        let center_x = size_info.width / 2.;
+        let center_y = size_info.height / 2.;
+        let x = (rect.x - center_x) / center_x;
+        let y = -(rect.y - center_y) / center_y;
+        let width = rect.width / center_x;
+        let height = rect.height / center_y;
+        let color = Rgba {
+            r: rect.color.r,
+            g: rect.color.g,
+            b: rect.color.b,
+            a: (rect.alpha * 255.) as u8,
+        };
+
+        self.vertices.push(Vertex { x, y: y + height, color });
+        self.vertices.push(Vertex { x, y, color });
+        self.vertices.push(Vertex { x: x + width, y: y + height, color });
+        self.vertices.push(Vertex { x: x + width, y, color });
+
+        self.indices.push(index);
+        self.indices.push(index + 1);
+        self.indices.push(index + 2);
+
+        self.indices.push(index + 2);
+        self.indices.push(index + 3);
+        self.indices.push(index + 1);
+
+        Ok(())
+    }
+
+    pub fn draw(&self) {
+        let size_info = self.size_info.as_ref().unwrap();
+        if self.indices.is_empty() {
+            return;
+        }
+
+        // Swap to rectangle rendering program.
+        unsafe {
+            // Swap program.
+            gl::UseProgram(self.program.id);
+
+            // Remove padding from viewport.
+            gl::Viewport(0, 0, size_info.width as i32, size_info.height as i32);
+
+            // Change blending strategy.
+            gl::Enable(gl::BLEND);
+            gl::BlendFuncSeparate(gl::SRC_ALPHA, gl::ONE_MINUS_SRC_ALPHA, gl::SRC_ALPHA, gl::ONE);
+
+            // Setup data and buffers.
+            gl::BindVertexArray(self.vao);
+
+            gl::BindBuffer(gl::ELEMENT_ARRAY_BUFFER, self.ebo);
+            gl::BufferData(
+                gl::ELEMENT_ARRAY_BUFFER,
+                (self.indices.len() * std::mem::size_of::<u16>()) as isize,
+                self.indices.as_ptr() as *const _,
+                gl::STREAM_DRAW,
+            );
+
+            gl::BindBuffer(gl::ARRAY_BUFFER, self.vbo);
+            gl::BufferData(
+                gl::ARRAY_BUFFER,
+                (self.vertices.len() * std::mem::size_of::<Vertex>()) as isize,
+                self.vertices.as_ptr() as *const _,
+                gl::STREAM_DRAW,
+            );
+
+            // Position.
+            gl::VertexAttribPointer(
+                0,
+                2,
+                gl::FLOAT,
+                gl::FALSE,
+                (size_of::<Vertex>()) as _,
+                ptr::null(),
+            );
+            gl::EnableVertexAttribArray(0);
+
+            // Color
+            gl::VertexAttribPointer(
+                1,
+                4,
+                gl::UNSIGNED_BYTE,
+                gl::TRUE,
+                (size_of::<Vertex>()) as _,
+                offset_of!(Vertex, color) as *const _,
+            );
+            gl::EnableVertexAttribArray(1);
+
+            gl::DrawElements(
+                gl::TRIANGLES,
+                self.indices.len() as i32,
+                gl::UNSIGNED_SHORT,
+                ptr::null(),
+            );
+
+            // Deactivate rectangle program again.
+            // Reset blending strategy.
+            gl::Disable(gl::BLEND);
+            gl::BlendFunc(gl::SRC_COLOR, gl::ONE_MINUS_SRC_COLOR);
+
+            // Reset data and buffers.
+            gl::BindBuffer(gl::ARRAY_BUFFER, 0);
+            gl::BindVertexArray(0);
+
+            let padding_x = size_info.padding_x as i32;
+            let padding_y = size_info.padding_y as i32;
+            let width = size_info.width as i32;
+            let height = size_info.height as i32;
+            gl::Viewport(padding_x, padding_y, width - 2 * padding_x, height - 2 * padding_y);
+
+            // Disable program.
+            gl::UseProgram(0);
+        }
+    }
+}
